+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## Next
 
+* New variable `company-lighter-base`.
 * Better tracking of the current selection.
 * Pressing `M-0`...`M-9` works in the search mode.
 * Pressing `<up>` or `<down>` doesn't quit the search mode.